requirements.txt
streamlit>=1.22.0
requests>=2.28.1
ollama>=0.1.2
python-dotenv>=0.21.0
pandas>=1.5.0